chore: rename from configuration to internal
